+++ packages/cms/lib/modules/resource-representation-widgets/lib/fields.js
@@ -19,7 +19,8 @@ const fields = [
           'shareChannelsSelection',
           'defaultImage',
           'hideStatus',
-          'shareTitle'
+          'shareTitle',
+          'multipleImages'
         ].concat(labels).concat(timeLabels)
       },
       {
@@ -145,7 +146,24 @@ const fields = [
     name: 'shareTitle',
     type: 'string',
     label: 'Title above share buttons (default: Deel dit voorstel)'
-  }
+  },
+  {
+    name: 'multipleImages',
+    type: 'boolean',
+    label: 'Display multiple images as a slider?',
+    choices: [
+          {
+              value: true,
+              label: "Yes",
+          },
+          {
+              value: false,
+              label: "No"
+          },
+      ],
+      def: false
+  },
+
 ].concat(
     ideaStates.map((state) => {
       return {
@@ -162,6 +180,38 @@ const fields = [
         name: 'labelTime' +  state.value,
         label: 'Labelfor time status: : ' + state.value,
       }
-    }));
+    })).concat(
+      [  
+        {
+          name: 'budgetTitle',
+          label: 'Budget: title for the budget header',
+          type: 'string',
+          placeholder: 'Budget',
+          def: 'Budget'
+        },
+        {
+        name: 'budgetFileOverviewFileTitle',
+        label: 'Budget-documents: alternative title for the "Bestanden" column',
+        type: 'string',
+        placeholder: 'Bestanden',
+        def: 'Bestanden'
+      },
+      {
+        name: 'budgetFileOverviewNameTitle',
+        label: 'Budget-documents: alternative title for the "Naam" column',
+        type: 'string',
+        placeholder: 'Naam',
+        def: 'Naam'
+      },
+    
+      {
+        name: 'budgetFileOverviewDateTitle',
+        label: 'Budget-documents: alternative title for the "Datum" column',
+        type: 'string',
+        placeholder: 'Datum',
+        def: 'Datum'
+      },
+    ]
+    );
 
 module.exports = fields;