+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## UNRELEASED
+* Add rich text editing functionality to the accordion widget
+* Allow arguments-widget and arguments-form-widget to be used outside of a resource page by setting the idea ID in the options.
+* Add functionality to display multiple images as a slider (with thumbnails) on the resource representation widget
+* Make image size configurable for the image widget
+* Add global setting to set openstad components cdn url
+
+
+
 ## 1.0.0
 * Add submission to resource form with configurable confirmation settings
 * Update ideas-on-map config to work with react-leaflet
@@ -71,8 +80,6 @@
 
 ## v0.20.3
 * Make sure ideas are existing on the page in Participatory budgetting
-
-## Unreleased
 * Add a new text style to select in the rich text editor.
 
 ## v0.20.2